Adds signal handling to track monthly active users (MAU)
This commit adds signal handling to create a record in
figures.models.CourseMonthlyActiveUser model, using the
CourseMonthlyActiveUserManager.add_mau method.

This commit also logs the event to the tracking logs for initial
debugging

Of note is that the signal does not have a reference to the request, so
it needs to get the request object from cache, using `crum`.

Also, we need to test this in both multisite and standalone so that the
site object can be accurately referenced as a foreign key relationship
to the MAU event